refactor(webpack): split app-shell webpack into chunks
