+++ apps/AppShell/config/webpack/baseLoaders.js
@@ -1,0 +1,101 @@
+const getClientEnvironment = require('../env')
+const paths = require('../paths')
+
+const hasJsxRuntime = (() => {
+  if (process.env.DISABLE_NEW_JSX_TRANSFORM === 'true') {
+    return false
+  }
+
+  try {
+    require.resolve('react/jsx-runtime')
+    return true
+  } catch (e) {
+    return false
+  }
+})()
+
+const baseLoaders = (webpackEnv, appEnv) => {
+  const isEnvProduction = webpackEnv === 'production'
+  const isEnvDevelopment = webpackEnv === 'development'
+  const isEnvServer = appEnv === 'server'
+  const isEnvClient = appEnv === 'client'
+
+  // We will provide `paths.publicUrlOrPath` to our app
+  // as %PUBLIC_URL% in `index.html` and `process.env.PUBLIC_URL` in JavaScript.
+  // Omit trailing slash as %PUBLIC_URL%/xyz looks better than %PUBLIC_URL%xyz.
+  // Get environment variables to inject into our app.
+  const env = getClientEnvironment(paths.publicUrlOrPath.slice(0, -1))
+
+  const shouldUseReactRefresh = env.raw.FAST_REFRESH
+
+  return [
+    // Process application JS with Babel.
+    // The preset includes JSX, Flow, TypeScript, and some ESnext features.
+    {
+      test: /\.(js|mjs|jsx|ts|tsx)$/,
+      include: paths.appSrc,
+      loader: require.resolve('babel-loader'),
+      options: {
+        customize: require.resolve('babel-preset-react-app/webpack-overrides'),
+
+        presets: isEnvClient
+          ? [
+              [
+                require.resolve('babel-preset-react-app'),
+                {
+                  runtime: hasJsxRuntime ? 'automatic' : 'classic',
+                },
+              ],
+            ]
+          : undefined,
+
+        plugins: [
+          // Add support for styled-components ssr
+          require.resolve('babel-plugin-styled-components'),
+          // Transform dynamic import to require for server
+          isEnvServer && require.resolve('babel-plugin-dynamic-import-node'),
+          [
+            require.resolve('babel-plugin-named-asset-import'),
+            {
+              loaderMap: {
+                svg: {
+                  ReactComponent: '@svgr/webpack?-svgo,+titleProp,+ref![path]',
+                },
+              },
+            },
+          ],
+          isEnvClient && isEnvDevelopment && shouldUseReactRefresh && require.resolve('react-refresh/babel'),
+        ].filter(Boolean),
+        // This is a feature of `babel-loader` for webpack (not Babel itself).
+        // It enables caching results in ./node_modules/.cache/babel-loader/
+        // directory for faster rebuilds.
+        cacheDirectory: true,
+        cacheCompression: isEnvProduction,
+        compact: isEnvProduction,
+      },
+    },
+    // Process any JS outside of the app with Babel.
+    // Unlike the application JS, we only compile the standard ES features.
+    {
+      test: /\.(js|mjs)$/,
+      exclude: /@babel(?:\/|\\{1,2})runtime/,
+      loader: require.resolve('babel-loader'),
+      options: {
+        babelrc: false,
+        configFile: false,
+        compact: false,
+        presets: [[require.resolve('babel-preset-react-app/dependencies'), {helpers: true}]],
+        cacheDirectory: true,
+        cacheCompression: isEnvProduction,
+
+        // If an error happens in a package, it's possible to be
+        // because it was compiled. Thus, we don't want the browser
+        // debugger to show the original code. Instead, the code
+        // being evaluated would be much more helpful.
+        sourceMaps: false,
+      },
+    },
+  ]
+}
+
+module.exports = {baseLoaders}
