+++ apps/AppShell/config/webpack/client.js
@@ -1,14 +1,58 @@
-// const webpack = require('webpack')
-const nodeExternals = require('webpack-node-externals')
+const fs = require('fs')
+const path = require('path')
 
-// const getClientEnvironment = require('../env')
-// const paths = require('../paths')
+const CssMinimizerPlugin = require('css-minimizer-webpack-plugin')
+const ModuleScopePlugin = require('react-dev-utils/ModuleScopePlugin')
+const TerserPlugin = require('terser-webpack-plugin')
+const {merge} = require('webpack-merge')
+const nodeExternals = require('webpack-node-externals')
+// const nodeExternals = require('webpack-node-externals')
+
+const {version} = require('../../package.json')
+const getClientEnvironment = require('../env')
+const modules = require('../modules')
+const paths = require('../paths')
 const getBaseConfig = require('./base')
+const {clientLoaders} = require('./clientLoaders')
+const {clientPlugins} = require('./clientPlugins')
+const createEnvironmentHash = require('./persistentCache/createEnvironmentHash')
+
+// const {baseLoaders} = require('./baseLoaders')
+// const packageJsonDeps = require('../../package.json').peerDependencies
+
+// Source maps are resource heavy and can cause out of memory issue for large source files.
+const shouldUseSourceMap = process.env.GENERATE_SOURCEMAP !== 'false'
+
+const reactRefreshRuntimeEntry = require.resolve('react-refresh/runtime')
+const reactRefreshWebpackPluginRuntimeEntry = require.resolve('@pmmmwh/react-refresh-webpack-plugin')
+const babelRuntimeEntry = require.resolve('babel-preset-react-app')
+const babelRuntimeEntryHelpers = require.resolve('@babel/runtime/helpers/esm/assertThisInitialized', {
+  paths: [babelRuntimeEntry],
+})
+const babelRuntimeRegenerator = require.resolve('@babel/runtime/regenerator', {
+  paths: [babelRuntimeEntry],
+})
+
+// Check if TypeScript is setup
+const useTypeScript = fs.existsSync(paths.appTsConfig)
 
 // const {packageJsonDeps, version} = require('../../package.json')
 
 module.exports = webpackEnv => {
-  const baseConfig = getBaseConfig(webpackEnv)
+  const baseConfig = getBaseConfig(webpackEnv, 'client')
+
+  const isEnvDevelopment = webpackEnv === 'development'
+  const isEnvProduction = webpackEnv === 'production'
+
+  // Variable used for enabling profiling in Production
+  // passed into alias object. Uses a flag if passed into the build command
+  const isEnvProductionProfile = isEnvProduction && process.argv.includes('--profile')
+
+  // We will provide `paths.publicUrlOrPath` to our app
+  // as %PUBLIC_URL% in `index.html` and `process.env.PUBLIC_URL` in JavaScript.
+  // Omit trailing slash as %PUBLIC_URL%/xyz looks better than %PUBLIC_URL%xyz.
+  // Get environment variables to inject into our app.
+  const env = getClientEnvironment(paths.publicUrlOrPath.slice(0, -1))
 
   // const env = getClientEnvironment(paths.publicUrlOrPath.slice(0, -1))
 
@@ -20,11 +64,185 @@ module.exports = webpackEnv => {
   //   }))
   //   .reduce((acc, n) => ({...acc, ...n}), {})
 
-  return {
-    ...baseConfig,
+  const clientConfig = {
     target: 'web',
     externals: [nodeExternals()],
     externalsPresets: {node: true},
+    mode: isEnvProduction ? 'production' : isEnvDevelopment && 'development',
+    // Stop compilation early in production
+    bail: isEnvProduction,
+    // eslint-disable-next-line no-nested-ternary
+    devtool: isEnvProduction
+      ? shouldUseSourceMap
+        ? 'source-map'
+        : false
+      : isEnvDevelopment && 'cheap-module-source-map',
+    // These are the "entry points" to our application.
+    // This means they will be the "root" imports that are included in JS bundle.
+    entry: paths.appClientIndexTsx,
+    output: {
+      // The build folder.
+      path: isEnvDevelopment ? paths.appDistPublic : paths.appBuildPublic,
+      // Add /* filename */ comments to generated require()s in the output.
+      pathinfo: isEnvDevelopment,
+      // There will be one main bundle, and one file per asynchronous chunk.
+      // In development, it does not produce real files.
+      filename: isEnvProduction
+        ? `static/js/[name].${version}.[contenthash:8].js`
+        : isEnvDevelopment && 'static/js/bundle.js',
+      // There are also additional JS chunk files if you use code splitting.
+      chunkFilename: isEnvProduction
+        ? `static/js/[name].${version}.[contenthash:8].chunk.js`
+        : isEnvDevelopment && 'static/js/[name].chunk.js',
+      assetModuleFilename: 'static/media/[name].[contenthash:8][ext]',
+      // webpack uses `publicPath` to determine where the app is being served from.
+      // It requires a trailing slash, or the file assets will get an incorrect path.
+      // We inferred the "public path" (such as / or /my-project) from homepage.
+      // publicPath: paths.publicUrlOrPath,
+      publicPath: `${env.raw.ASSETS_PATH}/`,
+      // Point sourcemap entries to original disk location (format as URL on Windows)
+      devtoolModuleFilenameTemplate: isEnvProduction
+        ? info => path.relative(paths.appSrc, info.absoluteResourcePath).replace(/\\/g, '/')
+        : isEnvDevelopment && (info => path.resolve(info.absoluteResourcePath).replace(/\\/g, '/')),
+    },
+    cache: {
+      type: 'filesystem',
+      version: createEnvironmentHash(env.raw),
+      cacheDirectory: paths.appWebpackCache,
+      store: 'pack',
+      buildDependencies: {
+        defaultWebpack: ['webpack/lib/'],
+        config: [__filename],
+        tsconfig: [paths.appTsConfig, paths.appJsConfig].filter(f => fs.existsSync(f)),
+      },
+    },
+    infrastructureLogging: {
+      level: 'none',
+    },
+    optimization: {
+      minimize: isEnvProduction,
+      minimizer: [
+        // This is only used in production mode
+        new TerserPlugin({
+          terserOptions: {
+            parse: {
+              // We want terser to parse ecma 8 code. However, we don't want it
+              // to apply any minification steps that turns valid ecma 5 code
+              // into invalid ecma 5 code. This is why the 'compress' and 'output'
+              // sections only apply transformations that are ecma 5 safe
+              // https://github.com/facebook/create-react-app/pull/4234
+              ecma: 8,
+            },
+            compress: {
+              ecma: 5,
+              warnings: false,
+              // Disabled because of an issue with Uglify breaking seemingly valid code:
+              // https://github.com/facebook/create-react-app/issues/2376
+              // Pending further investigation:
+              // https://github.com/mishoo/UglifyJS2/issues/2011
+              comparisons: false,
+              // Disabled because of an issue with Terser breaking valid code:
+              // https://github.com/facebook/create-react-app/issues/5250
+              // Pending further investigation:
+              // https://github.com/terser-js/terser/issues/120
+              inline: 2,
+            },
+            mangle: {
+              safari10: true,
+            },
+            // Added for profiling in devtools
+            keep_classnames: isEnvProductionProfile,
+            keep_fnames: isEnvProductionProfile,
+            output: {
+              ecma: 5,
+              comments: false,
+              // Turned on because emoji and regex is not minified properly using default
+              // https://github.com/facebook/create-react-app/issues/2488
+              ascii_only: true,
+            },
+          },
+        }),
+        // This is only used in production mode
+        new CssMinimizerPlugin(),
+      ],
+      splitChunks: isEnvProduction && {
+        chunks: 'all',
+        name: false,
+      },
+    },
+    resolve: {
+      // These are the reasonable defaults supported by the Node ecosystem.
+      // We also include JSX as a common component filename extension to support
+      // some tools, although we do not recommend using it, see:
+      // https://github.com/facebook/create-react-app/issues/290
+      // `web` extension prefixes have been added for better support
+      // for React Native Web.
+      extensions: paths.moduleFileExtensions.map(ext => `.${ext}`).filter(ext => useTypeScript || !ext.includes('ts')),
+      alias: {
+        // Support React Native Web
+        // https://www.smashingmagazine.com/2016/08/a-glimpse-into-the-future-with-react-native-for-web/
+        'react-native': 'react-native-web',
+        // Allows for better profiling with ReactDevTools
+        ...(isEnvProductionProfile && {
+          'react-dom$': 'react-dom/profiling',
+          'scheduler/tracing': 'scheduler/tracing-profiling',
+        }),
+        ...(modules.webpackAliases || {}),
+      },
+      plugins: [
+        // Prevents users from importing files from outside of src/ (or node_modules/).
+        // This often causes confusion because we only process files within src/ with babel.
+        // To fix this, we prevent you from importing files out of src/ -- if you'd like to,
+        // please link the files into your node_modules/ and let module-resolution kick in.
+        // Make sure your source files are compiled, as they will not be processed in any way.
+        new ModuleScopePlugin(paths.appSrc, [
+          paths.appPackageJson,
+          reactRefreshRuntimeEntry,
+          reactRefreshWebpackPluginRuntimeEntry,
+          babelRuntimeEntry,
+          babelRuntimeEntryHelpers,
+          babelRuntimeRegenerator,
+        ]),
+      ],
+    },
+    // TODO: Make base config load the baseLoaders, and not inside the client and server
+    // module: {
+    //   strictExportPresence: true,
+    //   rules: [
+    //     {
+    //       // "oneOf" will traverse all following loaders until one will
+    //       // match the requirements. When no loader matches it will fall
+    //       // back to the "file" loader at the end of the loader list.
+    //       oneOf: [
+    //         ...baseLoaders(webpackEnv, 'client'),
+    //       ],
+    //     },
+    //   ].filter(Boolean),
+    // },
+    plugins: [...clientPlugins(webpackEnv)].filter(Boolean),
+    // Turn off performance processing because we utilize
+    // our own hints via the FileSizeReporter
+    performance: false,
+    module: {
+      strictExportPresence: true,
+      rules: [
+        // Handle node_modules packages that contain sourcemaps
+        shouldUseSourceMap
+          ? {
+              enforce: 'pre',
+              exclude: /@babel(?:\/|\\{1,2})runtime/,
+              test: /\.(js|mjs|jsx|ts|tsx|css)$/,
+              loader: require.resolve('source-map-loader'),
+            }
+          : {},
+        {
+          // "oneOf" will traverse all following loaders until one will
+          // match the requirements. When no loader matches it will fall
+          // back to the "file" loader at the end of the loader list.
+          oneOf: [...clientLoaders(webpackEnv)].filter(Boolean),
+        },
+      ],
+    },
     // plugins: baseConfig.plugins.concat([
     //   new webpack.container.ModuleFederationPlugin({
     //     name: 'webpackHost',
@@ -45,4 +263,6 @@ module.exports = webpackEnv => {
     //   }),
     // ]),
   }
+
+  return merge(baseConfig, clientConfig)
 }
