+++ apps/AppShell/config/webpack/serverLoaders.js
@@ -3,8 +3,6 @@ const postCssPresetEnv = require('postcss-preset-env')
 // eslint-disable-next-line import/no-extraneous-dependencies
 const getCSSModuleLocalIdent = require('react-ssr-dev-utils/getCSSModuleLocalIdent')
 
-const paths = require('../paths')
-
 // Source maps are resource heavy and can cause out of memory issue for large source files.
 const shouldUseSourceMap = process.env.GENERATE_SOURCEMAP !== 'false'
 
@@ -15,68 +13,6 @@ const cssRegex = /\.css$/
 const cssModuleRegex = /\.module\.css$/
 const sassRegex = /\.(scss|sass)$/
 const sassModuleRegex = /\.module\.(scss|sass)$/
-
-const baseLoaders = (webpackEnv, appEnv) => {
-  const isEnvProduction = webpackEnv === 'production'
-  const isEnvServer = appEnv === 'server'
-
-  return [
-    // Process application JS with Babel.
-    // The preset includes JSX, Flow, TypeScript, and some ESnext features.
-    {
-      test: /\.(js|mjs|jsx|ts|tsx)$/,
-      include: paths.appSrc,
-      loader: require.resolve('babel-loader'),
-      options: {
-        customize: require.resolve('babel-preset-react-app/webpack-overrides'),
-
-        plugins: [
-          // Add support for styled-components ssr
-          require.resolve('babel-plugin-styled-components'),
-          // Transform dynamic import to require for server
-          isEnvServer && require.resolve('babel-plugin-dynamic-import-node'),
-          [
-            require.resolve('babel-plugin-named-asset-import'),
-            {
-              loaderMap: {
-                svg: {
-                  ReactComponent: '@svgr/webpack?-svgo,+titleProp,+ref![path]',
-                },
-              },
-            },
-          ],
-        ].filter(Boolean),
-        // This is a feature of `babel-loader` for webpack (not Babel itself).
-        // It enables caching results in ./node_modules/.cache/babel-loader/
-        // directory for faster rebuilds.
-        cacheDirectory: true,
-        cacheCompression: isEnvProduction,
-        compact: isEnvProduction,
-      },
-    },
-    // Process any JS outside of the app with Babel.
-    // Unlike the application JS, we only compile the standard ES features.
-    {
-      test: /\.(js|mjs)$/,
-      exclude: /@babel(?:\/|\\{1,2})runtime/,
-      loader: require.resolve('babel-loader'),
-      options: {
-        babelrc: false,
-        configFile: false,
-        compact: false,
-        presets: [[require.resolve('babel-preset-react-app/dependencies'), {helpers: true}]],
-        cacheDirectory: true,
-        cacheCompression: isEnvProduction,
-
-        // If an error happens in a package, it's possible to be
-        // because it was compiled. Thus, we don't want the browser
-        // debugger to show the original code. Instead, the code
-        // being evaluated would be much more helpful.
-        sourceMaps: false,
-      },
-    },
-  ]
-}
 
 const serverLoaders = webpackEnv => {
   const isEnvProduction = webpackEnv === 'production'
@@ -178,4 +114,4 @@ const serverLoaders = webpackEnv => {
   ]
 }
 
-module.exports = {baseLoaders, serverLoaders}
+module.exports = {serverLoaders}
