+++ package.json
@@ -10,11 +10,11 @@
     "docs": "npm run docs -ws --if-present",
     "prepare": "husky install",
     "lint-staged": "lint-staged",
-    "app-shell:lint": "npm run lint -w @next/app-shell --if-present", 
-    "app-shell:build": "npm run build -w @next/app-shell --if-present", 
-    "app-shell:test": "npm run test -w @next/app-shell --if-present", 
-    "app-shell:serve": "npm run serve -w @next/app-shell --if-present",  
-    "app-shell:dev": "npm run dev -w @next/app-shell --if-present", 
+    "app-shell:lint": "npm run lint -w @next/app-shell --if-present",
+    "app-shell:build": "npm run build -w @next/app-shell --if-present",
+    "app-shell:test": "npm run test -w @next/app-shell --if-present",
+    "app-shell:serve": "npm run serve -w @next/app-shell --if-present",
+    "app-shell:dev": "npm run dev -w @next/app-shell --if-present",
     "app-shell:ratchet": "npm run jest-ratchet -w @next/app-shell --if-present",
     "header:build": "npm run build -w @next/header --if-present",
     "header:lint": "npm run lint -w @next/header --if-present",
